+++ packages/connectanum_mcp/lib/connectanum_mcp.dart
@@ -1,5 +1,7 @@
 library;
 
+export 'package:connectanum_core/connectanum_core.dart'
+    show McpResourceUriTemplate;
 export 'src/protocol/capabilities.dart';
 export 'src/protocol/constants.dart';
 export 'src/protocol/errors.dart';